Replace children in ArticleParamsForm
